pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vyasa"
version = "0.3.47"
description = "A lightweight, elegant blogging platform built with FastHTML"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Yeshwanth"},
]
keywords = ["fasthtml", "blog", "markdown", "htmx", "mermaid", "sidenotes"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "python-fasthtml>=0.6.9",
    "mistletoe>=1.4.0",
    "python-frontmatter>=1.1.0",
    "uvicorn>=0.30.0",
    "monsterui>=0.0.37",
    "pylogue>=0.3",
    "loguru>=0.7.3",
    "fastsql>=2.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
]
auth = [
    "authlib>=1.3.0",
]

[project.scripts]
vyasa = "vyasa.main:cli"

[project.urls]
Homepage = "https://github.com/yeshwanth/vyasa"
Repository = "https://github.com/yeshwanth/vyasa"
Issues = "https://github.com/yeshwanth/vyasa/issues"

[tool.setuptools]
packages = ["vyasa"]
include-package-data = true

[tool.setuptools.package-data]
vyasa = ["static/*", "static/**/*"]

[tool.uv]
cache-keys = [{file = "pyproject.toml"}]
